requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
datasets==3.6.0
dill==0.3.8
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==2.3.3
flask-cors==6.0.1
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gTTS==2.5.4
gunicorn
hf-xet==1.1.5
huggingface-hub==0.33.0
HyperPyYAML==1.2.2
idna==3.10
ipinfo==5.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.0
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.9
pyarrow==20.0.0
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
speechbrain==1.0.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.52.4
vosk==0.3.44 # changed from 0.3.45
typing_extensions==4.11.0
